move fitswcs and sliced_wcs converters to wcs submodule
