+++ asdf_astropy/converters/fits/__init__.py
@@ -2,8 +2,6 @@ __all__ = [
     "FitsConverter",
     "AsdfFitsConverter",
     "AstropyFitsConverter",
-    "FitsWCSConverter",
 ]
 
 from .fits import AsdfFitsConverter, AstropyFitsConverter, FitsConverter
-from .fitswcs import FitsWCSConverter
