+++ asdf_astropy/converters/wcs/__init__.py
@@ -1,5 +1,6 @@
 __all__ = [
+    "FitsWCSConverter",
     "SlicedWCSConverter",
 ]
-
+from .fitswcs import FitsWCSConverter
 from .slicedwcs import SlicedWCSConverter
